requirements.txt
pypdf
python-dotenv
pdf2image
ultralytics
rapidocr-onnxruntime
pypdfium2
opencv-python-headless
pillow
pdfplumber
